requirements.txt
africastalking==1.2.9
alembic==1.16.3
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses==0.6
deprecation==2.1.0
dnspython==2.7.0
email-validator==2.1.0.post1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.5
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
postgrest==1.1.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
redis==5.2.1
flask-caching==2.3.0
celery==5.4.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
schema==0.7.7
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets>=11,<16
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn
intuit-oauth==1.2.4
python-quickbooks==0.9.5
requests-oauthlib==1.3.1
